Increase the -l option accuracy by biasing by the select() overhead time
While only intending to be approximate, the -l option for the CPU
loading wasn't as very accurate for high values of -l because of the
delay caused by select system call for low values of a delay. So
determine the overhead and bias the delay on the next iteration to
try and make this a little more accurate.

The following results were based on a 60 second run per -l option used
measured on an i5-3210M:

-l	measure
value	CPU load %
1	1
10	9.82
20	19.86
30	29.94
40	39.88
50	49.84
60	59.94
70	69.9
80	80.08
90	89.92
99	99.02
100	99.86

..so this minor fix gets fairly close to the desired -l value.

Thanks to Carsten Grohmann for bringing this to my attention.

Signed-off-by: Colin Ian King <colin.king@canonical.com>