Initialise finish time to be start time in case it is never run
We can have a situation where many stressors are invoked and the
timeout kicks in before they are all successfully started. In these
cases it can be that the finish time has not been set because the
process is not terminated (because it never started) and so we get
weird run time stats. So initialise the finish time to be the start
time and this can be set correctly when the process terminates later
on.

Signed-off-by: Colin Ian King <colin.king@canonical.com>