stress-fallocate: fix incorrect errors reported on fallocate sizes
The linux fallocate should truncate the file so that it is not larger
than the size of the next posix_fallocate.  When this occurred, the
stressor reported the file size was larger than exected, which was
a false error message. We need to truncate the file before moving
onto the posix_fallocate iteration to avoid this false error message.

Signed-off-by: Colin Ian King <colin.king@canonical.com>