stress-hdd: fix false error reports on short reads
Ensure we don't attempt to read off the end of the file and
get short reads.  First, we must make the file size at least the size
of the user specified write size. Secondly, ensure random read offsets
are in range so that we don't partially read off the end of the file.

Signed-off-by: Colin Ian King <colin.king@canonical.com>