sched: fix default priority levels and lack of class name in messages
default priority levels should be automatically set correctly when in
aggressive mode. Currently, if the user omits the sched-prio then the
scheduling settings exit stress-ng with an error, which is not useful
at all.

Instead, default the sched priority to be maximum when in aggressive
mode, and midway otherwise if the priority is not given.

If the priority is given non fifo/rr scheduling classes then just
set this to zero and issue a warning rather than stopping stress-ng
on an error.

Finally, the user won't be able to figure out what the scheduling
class is from the numeric values, so translate these to a human
readable string.

Also cater for deadline class on system where SCHED_DEADLINE is
defined.

Signed-off-by: Colin Ian King <colin.king@canonical.com>