stress-hdd: fix verify mode for hdd stressor
Somehow the verify mode got overlooked and the hdd stressor does not
sanity check what it wrote. This fix ensures data is always read back
and if the verify mode is enabled it will check that specific parts
of the data match a very simple 1 byte known value.  While this is a
naive check, it is better than no check.  The verify works for different
read/write modes, and has been tested with:

	seq read, seq write
	rnd read, seq write
	seq read, rnd write
	rnd read, rnd write

..with various file systems and buffer sizes too.

Signed-off-by: Colin Ian King <colin.king@canonical.com>