stress-ng: avoid builtins when libs shall be tested
The documentation of string and wide character string tests states it shall
test the glibc functions. But certain string functions are in gcc as builtin.

Thereby stress-ng was actually testing the compiler with some of these tests.
Adding -fno-builtin when compiling the respective tests makes sure it calls
the library as intended.

Signed-off-by: Christian Ehrhardt <paelzer@gmail.com>
Signed-off-by: Colin Ian King <colin.king@canonical.com>