stress-timer, itimer: fix termination hangs (LP: #1494751)
On slow devices with real time timers we can get stuck
handling timer signals and never get around to the main
loop that exits. Also, SIGALRM alarm and setitimer share
the same alarm, so we never get the termination alarm
and the test never exits.

Fix in the timer handler by:

1. checking for SIGINT explicity
2. periodic checking of timout

Signed-off-by: Colin Ian King <colin.king@canonical.com>