stress-ureadahead: provide feed back when timout occurs (LP: #1494784)
The readahead stressor can be interrupted because of a timeout
(e.g. if the timeout is too short and the temp data file is too
large on a slow I/O device) before the file is completed. Currently
this stressor aborts with no message indicated what has gone wrong.

Instead, the stress test needs to explain that the file was not
written out before the test duration limit was reached and we should
not abort with a failure message since the test technically didn't run.

Signed-off-by: Colin Ian King <colin.king@canonical.com>