stress-sigsuspend: remove spinlock to remove contention (LP: #1496439)
The spinlock was causing contention on large multi processor nodes
and also possible lock up if one of the child processes was killed
before the unlock.  So replace this with shared memory counters that
are spaced at a reasonable memory stride.  This allows us to crank
up the sigsuspend rate and also get some interesting results on NUMA
systems.

Signed-off-by: Colin Ian King <colin.king@canonical.com>