stress-fifo: avoid blocking on reader to avoid deadlock (LP: #1497268)
the read fifo child processes can get blocked if the parent is
killed and the fifo has not been opened by the writer parent.
The best way forward is for the child to open the fifo in a
non-blocking mode and select() on the file descriptor waiting
for data to arrive. We can then add a timeout to this and bail
out if need be.

Signed-off-by: Colin Ian King <colin.king@canonical.com>