stress-sem: handle signal during setup phase correctly
On slow systems or when hundreds of sem stressor child processes
are started it is possible to get a SIGALRM or SIGINT during the
setup phase and this is not detected. The code then blocks forever
paused waiting for the already occurred signal.  Fix this by
detecting if opt_do_run has changed state during the stressor
initialisation and reaping existing processes.

Also don't pause waiting explicitly for a signal, instead do
a 10Hz poll on opt_do_run and on the ops count.  Although this
is sub-optimal, it does allow us to work around a race window
between the end of the init and the waiting for a signal.

Signed-off-by: Colin Ian King <colin.king@canonical.com>