stress-shm-sysv: make shm cleanup robust
On small memory platforms where memory is low, the OOM killer can kill the
shm stressor resulting in left over unremoved shm segments.  Move the shm
allocation to a child processes that informs the parent of the shm ids so that
the parent can attempt to reap these if necessary.

Signed-off-by: Colin Ian King <colin.king@canonical.com>