stress-shm-sysv: add more allocation smarts
When running with --maximize, the System V shared memory stressor
can easily hit overcommit and the OOM killer can kill all stress-ng
processes.  To make the stressor less OOMy, smarts have been added
to guesstimate the maximum memory available before allocation (which
is racy, but better than nothing). Also, add a SIGALRM handler to the
child so that it can be signalled by the parent to abort and reap
share segments more effectively, rather than being killed and leaving
the parent to try and reap, which can miss segments that have been
created and not comminicated to the parernt over the messaging pipe.

Signed-off-by: Colin Ian King <colin.king@canonical.com>