stress-readahead: handle -ENOSPC as a non-failure condition
When a filesystem is full don't treat this as a failed test. Instead
get the size of the file, round it to the read size and read just
the partial file back.  This way on small systems we don't trip over
ENOSPC errors but instead maximize on what space we do have.

Signed-off-by: Colin Ian King <colin.king@canonical.com>