stress-dentry: fix file striding on clean-up (LP: #1547004)
using a prime stride is a good mechanism to step around files
in a directory in a non-sequential manner, however, the
implementation is broken because the large prime being used
to do the striding wraps on 64 bit integers causing aliasing
and hence not hitting all files.

Use a smaller prime and ensure that the number of files is not
a multiple of this prime to ensure the striding always works
correctly and allows us to remove all the files.

Signed-off-by: Colin Ian King <colin.king@canonical.com>