stress-mmap: handle SIGBUS signals (LP: #1569468)
We can have the possibility that the file backed memory mapping is
over committed and we don't actually have enough file backed pages
available, causing accesses to these mappings to trip a SIGBUS.
Handle these with a signal handler and a trampoline back to the
main stressor loop which aborts the rest of the page stressing and
re-tries.  Although this is not a perfect solution, it allows the
stressor to iterate over different mapping options and at least
exhaustively stress these options rather than bailing out early.

Signed-off-by: Colin Ian King <colin.king@canonical.com>