Ensure processes are kept to the CPU affinity set when stress-ng is invoked
Rather than allowing stressors to change affinity to just any of the
available CPUs we should constrain them to the CPU affinity set when
stress-ng was invoked.  This allows us to invoke stress-ng using tools
such as taskset and stress-ng will only change process affinity of
the stressors to the given set.

For example:

taskset -c 4-7 stress-ng --cache 4 --stream 4 --yield 4

..will keep all the stressors on CPUs 4,5,6 and 7.

Kudos to Jim Rowan <jim.rowan@intel.com> for the original idea and
example patch from which this patch was derived.

Signed-off-by: Colin Ian King <colin.king@canonical.com>