stress-icache: don't assume that read/write/exec protection is allowed
Some architectures and/or kernel configurations may not allow
READ/WRITE/EXEC protection on a text segment, so we need to cater for
this is a more generic way by changing page protection to READ | WRITE,
then modify the text segment, then change it back to READ | EXEC to
make this work across all supported achitectures.

Also, madvise the page not to be a huge page and don't run the stressor
if the page size is not 4K since we are aligning the page to a 4K
boundary and hence we need a minimum of 4K alignment.

These changes mean we now have less I$ ops per second, however, we
now do hit the TLB harder, which is an interesting plus point to make
this stressor a more useful CPU cache/TLB stress test.

Signed-off-by: Colin Ian King <colin.king@canonical.com>