stress-hsearch: don't free keys, some *BSD systems free them in hdelete
hdelete may or may not delete the key data passed into the hash, so it
is hard to know what is the right strategy when cleaning up.  If we
free the key data, it may break the malloc heap on OpenBSD; if we don't
free then we leak.  For now, make it leaky rather than break the heap.

Signed-off-by: Colin Ian King <colin.king@canonical.com>