Pass args to stressors via a args_t struct
This is the first step in cleaning up the stress-ng interfaces.
Instead of passing 4 args to each stressor, we instead pass
a struct populated with the args. This allows us to add extra
state info (such as pid, ppid, etc) of each stressor process
without the need to radically change the stressor interfaces.

Signed-off-by: Colin Ian King <colin.king@canonical.com>