stress-icache: cater for 4K, 16K and 64K page sizes and alignments
Cater for 4K, 16K and 64K page alignments. Thanks to Rob Colclaser
at Cavium for the idea and initial patch. This is re-working of
the patch from Rob using macros to do all the heavy lifting of very
similar functions that handle the hard coded page sizes and page
alignment magic.

Signed-off-by: Colin Ian King <colin.king@canonical.com>