stress-ng: fix rounding error in bogo ops-calculation
There is a somewhat non-intuitive operation if one specifies the
number of stressors and the number of bogo-ops on the command line
and the number of bogo-ops is less than the number of stressors.
The expectation in this case (#stressors > #bogo-ops) that one  would
effectively run one op on each stressor.  However, what happens is
the number of operations rounds to 0 and the test runs until the
timeout period is reached instead of the expected short test run time.

If we round up the bogo ops calculation, then we get 1 ops run in
this case which feels more like the expected behavior.

Signed-off-by: Rob Colclaser <Rob.Colclaser@cavium.com>
Signed-off-by: Colin Ian King <colin.king@canonical.com>