stress-vm: check for bit errors and report these rather than silently ignoring them
The vm stressor is not returning an EXIT_FAILURE if it detects bit
errors.  Add in accounting for this and report errors as well as returning
EXIT_FAILURE.

Many thanks to Rob Colclaser at Cavium for spotting this error and sending
me an intitial fix that was re-worked a little to produce this final fix.

Reported-by: Rob Colclaser <Rob.Colclaser@cavium.com>
Signed-off-by: Colin Ian King <colin.king@canonical.com>