stress-zlib: add check sum sanity checking and --zlib-method
This adds fixes an issue in the deflate and adds checksum sanity
checking on the data by comparing checksums of pre-deflated and
post-inflated data.  Also a --zlib-method option has been added
to allow selection of the way data to be deflated is generated.

Many thanks to Rob Colclaser at Cavium for the original patch
that was re-worked a little to add some SIGPIPE error handling
to produce this final patch.

Reported-by: Rob Colclaser <Rob.Colclaser@cavium.com>
Signed-off-by: Rob Colclaser <Rob.Colclaser@cavium.com>
Signed-off-by: Colin Ian King <colin.king@canonical.com>