stress_sighandler: ensure signal handlers use alternative signal stack
To be totally safe, ensure that when a signal handler is installed
we use a pre-allocated stactic mlock'd alternative signal stack.

This fixes an obscure bug in the stress-stackmmp stressor where
SIGALRM was causing problems handling a signal when the stack was
out of space and an SIGFAULT was being generated.

Note that we can get away with a static alternative stack as this
is a per-process stack allocation and this allows us to add a MLOCKED
attribute to try and ensure that the stack is always in memory so
we don't incur any page faulting VM issues when the VM is really
low on memory.

Kudos to Joseph DeVincentis at Cavlium for debugging the failures
in stress-stackmap and figuring out how to fix this issue.

Reported-by: Joseph DeVincentis <Joseph.DeVincentis@cavium.com>
Signed-off-by: Colin Ian King <colin.king@canonical.com>