Replace procs[] array with a doubly-linked list of proc_info_t
The procs[] array is cumbersome, so this major re-work replaces it
with a doubly-linked list.  By also referencing the parent stressor
to each proc_info_t item we can also optimize the way we look up
information too.  The size of the stat information buffer is also
pointed to by the proc_info_t allowing a simpler lookup of each
proc_info's stats data. The end result is a better packed shared
data section too.

Signed-off-by: Colin Ian King <colin.king@canonical.com>