stress-softlockup: try to synchronize start of all children
If we are over-committed with more stressors than online CPUs we
end up with some stress workers blocked and only start when a CPU
becomes free to run the next worker. Instead, ensure workers have
all got their child treads up and running with the same notion of
start time so that they all complete roughly at the same time
even if that means some terminate without being run for any length
of time because they have been blocked on the run queue.

Signed-off-by: Colin Ian King <colin.king@canonical.com>