stress-vforkmany: allow children to be OOM'd (LP: #1698747)
Ensure child processes can be OOM'd so that this stressor can
keep running.  Currently the parent of stress-ng can be OOM'd before
the child processes causing test frameworks or the user's shell or
login session to be OOM'd before the vfork processes. Set the OOM
adjust to preferentially OOM the children first.

Signed-off-by: Colin Ian King <colin.king@canonical.com>