stress-socket-fd: don't force overcommit on fd allocations (LP: #1692668)
When running as root, each socket-fd stressor can allocate a huge amount
of file descriptors which results in low memory situations where even
the OOM-killer can't find any more free pages.  This is a pathologically
bad situation and is one of the limits of the default overcommit memory
allocation policy of Linux.  Fix this for the root user condition by
finding the system limit of file descriptors, taking off a bit of slack
and dividing that by the number of instances.

Signed-off-by: Colin Ian King <colin.king@canonical.com>