stress-timer: ensure we stop timers when timer handler is saturated
When a system is totally saturated we never get around to checking if
we've hit the maximum number of bogo ops.  Copy the maximum ops to
a global variable so it can be accessed by the signal handler and
add some extra checking to ensure we've not overrun the maximum ops
inside the signal handler.

Reported by: Maxime CHEVALLIER <maxime.chevallier@smile.fr>
Signed-off-by: Colin Ian King <colin.king@canonical.com>