stress-inode-flags: reset flags to zero at end of each loop (LP: #1730297)
The multi-threaded setting of flags can mean that unsetting of the
immutable flag is racy and hence at the end of the stress iteration
we may have not cleared it. Hard force all the flags to zero at the
end of each iteration and before we close a file to ensure we don't
have the immutable bit set which stops us from removing the temp file.

Signed-off-by: Colin Ian King <colin.king@canonical.com>