stress-tlb: fix incorrect affinity on cores
This fixes an issue found and fixed by Rob Colclaser:

"...I was only seeing one core doing anything.  I started adding a bunch
of prints and I believe there is a bug with the cpu affinity usage in
the test.  The first iteration of the loop sees N cores and creates N
child processes, each of which runs on a separate core.  But on the
second and subsequent iterations, the code only sees 1 core in the
proc_mask so all the children run on one core, which was a different
core from run to run.  The issus is that in the first loop, whatever
child runs last leaves the affinity set up for that core.  On the second
and subsequent loops, the code only thinks there is one core."

Kudos to Rob for tracking down the bug and the fix.

[ Colin King ]
I re-worked the aggressive TLB child process killing to only kicks into
action if a wait gets interrupted perhaps by SIGALRM and SIGINT so that
the stressor can terminate quickly on these signals.

Signed-off-by: Rob Colclaser <Rob.Colclaser@cavium.com>
Signed-off-by: Colin Ian King <colin.king@canonical.com>