Don't exit child sub-processes with exit() (LP: #1777564)
When a child sub-process in any stressor exits with exit()
then the new atexit handler traps this as an illegal return
from the stressor and throws an error report.  Avoid this
by exiting all stressor child sub-processes with _exit()
to skip the atexit handler.

Also fix up some exits that should be returns in stressors.

Signed-off-by: Colin Ian King <colin.king@canonical.com>