stress-socket-fd: add more yield points on fd recv and closing
We need more yield points as potentially we can have tens of thousands
of file descriptors open and being received from and we need to check
if we've been interrupted to do a timely close and exit of the client
and server.  This is especially noticeable on large systems with many
stressors running, lots of file descriptor resources and slower CPUs.

Signed-off-by: Colin Ian King <colin.king@canonical.com>