Ensure all fail log messages set the error to EXIT_FAILURE
If a stressor returns EXIT_SUCCESS we want that turned to
EXIT_FAILURE if any of the pr_fail messages have been emitted
during a stress test. Add a logging helper to check if a failure
message has been logged and if so, set the return code to
EXIT_FAILURE when required.

Signed-off-by: Colin Ian King <colin.king@canonical.com>