stress-personality: don't verify personality setting, it does not work
Setting the personality and checking it against what as set in
verify mode does not work as some kernels can't handle some personality
modes and silently fail on the personality setting without informing
the userspace that it's not fully set.  Disable verification.
Instead, add sanity check for fetching the personality setting.

Signed-off-by: Colin Ian King <colin.king@canonical.com>