stress-sysfs: don't nullify sysfs_path, set it to empty string and lock it
We hit issues one large CPU systems where sysfs_path was being nullified
to mark completion without locking it and some of the child threads
were dereferencing this null pointer and getting segfaulted. Fix this
by locking sysfs_path to avoid a race and setting it to an empty string
and checking for this instead of a NULL pointer to make it safer.

Signed-off-by: Colin Ian King <colin.king@canonical.com>