stress-dev: remove FIBMAP ioctl on device (LP: #1800609)
FIBMAP on a block device is plainly wrong, it should only be applied
to a file. The FIBMAP ioctl to a scsi block device is the same as the
SCSI_IOCTL_SEND_COMMAND ioctl that requires a scsi_ioctl_command
struct. As it stands, we are passing an int with stack garbage as
this struct which is passing garbage to the ioctl. The deprecated
SCSI ioctl creates a scsi_request from the garbage and this causes
the kernel to correctly warn with a usercopy_abort that the structure
is bogos.

Fix this by not calling FIBMAP on a raw SCSI device.

Signed-off-by: Colin Ian King <colin.king@canonical.com>