stress-affinity: fix set affinity setting across all stressors (LP: #1803322)
The stress-affinity stressor checks to see if sched_setaffinity works,
however, in doing so sets the CPU affinity to CPU 0 for all stressors
because this is performed at the start of stress-ng and other stressors
inherit this affinity setting. Fix this by getting and then setting
the same affinity.

Thanks to tonnylu for spotting this mistake and the bisection to detect
the buggy commit.

Signed-off-by: Colin Ian King <colin.king@canonical.com>