Add log message locking.
Log messages appear without any form of locking which is fine for
single line messages, but for multiple pr_* messages we can get
intermixed output from different concurrent stressors.  Add some
lockable logging:

	bool lock = false;

	pr_lock(&lock);
	pr_inf_lock(&lock, "here is a message\n");
	pr_inf_lock(&lock, "here is another message\n");
	pr_unlock(&lock);

The above locking scheme stops other pr_* messages being emitted
until the unlock.  Locking is based on flock on stdout, and if
locking fails (or is not supported) then we are back to the normal
unlocked mode of output.

Signed-off-by: Colin Ian King <colin.king@canonical.com>