stress-mmap: be less noisy on mmap failures and fix directory cleanup (LP: #1807732)
In the case where the system memory is completely depleted and mmap fails
currently the mmap stressor reports this a failure and spams the output.
Treat this as a info message rather than an error case as a low-memory
failed allocation is a valid test case.

Also, bump the number of allocation attempts up to 64K attempts.

Finally, the child should not return via the clean up path and remove
the tmp test directory as the parent does this and we get race situations
and an error message when either the child or the parent both try to
remove a directory that has been removed by one or the other.  Fix this
by making the child exit via _exit()

Signed-off-by: Colin Ian King <colin.king@canonical.com>