stress-numa: fix calculation of max nodes (LP: #1803468)
Previously we were getting an upper bound to the number of max nodes
instead of the actual number. That made migrate_pages() fail in some
devices due to having bits in the masks set beyond the maximum node
identifier.

We use now the count of bits in "Mems_allowed" to calculate the maximum
number of NUMA nodes. Fixes LP: #1803468.

Signed-off-by: Alfonso Sanchez-Beato <alfonso.sanchez-beato@canonical.com>
Signed-off-by: Colin Ian King <colin.king@canonical.com>