stress-numa: fix calculation of max nodes
Previously we were getting an upper bound to the number of max nodes
instead of the actual number. That made migrate_pages() fail in some
devices due to having bits in the masks set beyond the maximum node
identifier.

We use now the count of bits in "Mems_allowed" to calculate the maximum
number of NUMA nodes. Fixes LP: #1803468.