stress-*: remove the use of global opt flags for non-global stressor settings
There are a range of stressor settings that are using the global
opt flags when instead they should be per stressor related settings.
Remove the global flag use and use the set_setting/get_setting mechanism
to set/get these boolean flags.  This may be more awkward, but we
are low on global bit flags and these flags should use the normal
settings mechanism.

Signed-off-by: Colin Ian King <colin.king@canonical.com>