stress-ramfs: fix umounting races
Make mount/umount stress a child rather than a clone, this allows
better process control.  Force re-try on umount until it really is
umounted, this fixes a race where umount returns 0 but the umount
is still occurring.  Don't kill the child process as this needs
to cleanup on the mount points.

Signed-off-by: Colin Ian King <colin.king@canonical.com>