stress-mcontend: ensure mmap'd file is sync'd with zero data (LP: #1845011)
Write zero's to backing store of at least one page size and sync
it so that we can be 100% sure that mmap'ing the page into memory
is backed with data so we don't get BUS errors on some platforms
such as armhf Raspi2.

Also fix broken sanity check on 2nd mmap out of memory failure.

Signed-off-by: Colin Ian King <colin.king@canonical.com>