stress-af-alg: fix sockaddr algorithm type on bind()
The AF_ALG family's bind() handler in the kernel requires
that the 'struct sockaddr_alg.salg_type' field to be one
of 'aead', 'hash', 'rng, 'skcipher'.

This is because it requests to load a module based on it
(see linux/crypto/af_alg.c :: alg_bind()):

  request_module("algif-%s", sa->salg_type);

With the respective files being:

  $ ls -1 crypto/algif_*.c
  crypto/algif_aead.c
  crypto/algif_hash.c
  crypto/algif_rng.c
  crypto/algif_skcipher.c

Unfortunately the 'type' field in /proc/crypto does not match
the required values, and thus stress_af_alg_[hash|cipher]()
fail to bind() with ENOENT errno.
(stress_af_alg_rng() uses string "rng" so it's already good.)

Test-case: command to filter bind() syscalls exit/errno and
crypto algorithm name uniquely, and format it 'conveniently'
for comparison with 'diff' later (to check for what changes
happened) -- there should be a simpler way to do this.

  $ strace -e bind -f  ./stress-ng --af-alg 1 --timeout 5 2>&1 \
      | grep -o 'bind(.*) = .*' | sort -u | cut -d= -f3- \
      | sed 's,\\0, ,g' | tr -s ' '  | awk '{ print $2, $1, $6, $7 }' \
      | sort > $FILE

Before -- lots of ENOENTs and few 0s:

  $ cat $FILE | cut -d' ' -f3- | sort | uniq -c
     35 -1 ENOENT
      9 0

Ater -- most ENOENTs turned to EBUSYs:

  $ cat $FILE | cut -d' ' -f3- | sort | uniq -c
     35 -1 EBUSY
      9 -1 ENOENT
      1 0

It turned out this uncovered another regression, in the
stress_af_alg_hash() function, which caused some of the
'0' results to fail with EBUSY too.

This will be addressed in an upcoming patch. For now,
the problem with too many ENOENTs has been addressed.

Signed-off-by: Mauricio Faria de Oliveira <mfo@canonical.com>