stress-af-alg: fix close(fd) to avoid bind() EBUSY
The bind() EBUSY errno happen because the AF_ALG socket
has a reference being held; this became more evident on
the strace bind() output mentioned in previous patches:

  $ grep digest_null $FILE
  digest_null "hash -1 EBUSY
  digest_null "hash 0

This algorithm succeeded and failed; looking at strace
it succeeded first and later failed. And the kernel code
showed that the reason is the socket refcnt is not zero.

The recent refactor for /proc/crypto changed the behavior
of one graceful error-handling case for the send() syscall
in stress_af_alg_hash() in the "for (j = ..." loop.

Previously the loop just continue'd, and upon finishing it,
'close(fd)' was called.  After the refactor, the close(fd)
was brought in inside the loop (oversight? it breaks the loop), and
the first error handler returns EXIT_SUCCESS without close(fd),
this keeping the reference count greater than zero - which
caused the EBUSY errno on bind().

Address this by reverting the for-loop/error-handler to keep
the previous/existing behavior, which is just 'continue' and
close(fd) after the loop.

This change resolves all EBUSY failures:

  $ cat $FILE | cut -d' ' -f3- | sort | uniq -c
      9 -1 ENOENT
     35 0

Signed-off-by: Mauricio Faria de Oliveira <mfo@canonical.com>