stress-af-alg: skip 'internal' crypto algorithms
Some of the remaining bind() failures with ENOENT errno happen
due to attempting to use 'internal' cryptographic algorithms,
which should not happen (see linux/include/linux.h):

  /*
   * Mark a cipher as a service implementation only usable by another
   * cipher and never by a normal user of the kernel crypto API
   */
  #define CRYPTO_ALG_INTERNAL             0x00002000

Fortunately this information is available in /proc/crypto,
so parse that as well into 'struct crypto_info', and just
check/skip those in stress_af_alg().

Before -- many ENOENTs:

  $ cat $FILE | cut -d' ' -f3- | sort | uniq -c
     33 -1 ENOENT
     55 0

After -- just a few ENOENTs:

  $ cat $FILE | cut -d' ' -f3- | sort | uniq -c
      9 -1 ENOENT
     55 0

The remaining ENOENTs are apparently something specific to crypto
algorithms that are not found but require this 'cryptomgr' module
to be loaded (which is not available/happening).

I couldn't check in detail but kprobes traces show the notification
path returning NOTIFY_DONE (see linux/include/linux/notifier.h) after
attempting to request_module("cryptomgr") in crypto_probing_notify()
(invoked from the bind() path.)  For example:

	 stress-ng-af-al-26652 [048] .... 12159.926222: sys_bind(fd: 4, umyaddr: 7ffda8630420, addrlen: 58)
	 stress-ng-af-al-26652 [048] .... 12159.926230: p_crypto_alloc_skcipher_0: (crypto_alloc_skcipher+0x0/0x20) name="cast6"
	 stress-ng-af-al-26652 [048] .... 12159.926230: p_crypto_alg_mod_lookup_0: (crypto_alg_mod_lookup+0x0/0x210) name="cast6"
	 stress-ng-af-al-26652 [048] .... 12159.930060: p_crypto_probing_notify_0: (crypto_probing_notify+0x0/0x50)
	 stress-ng-af-al-26652 [048] d... 12159.931817: r_crypto_probing_notify_0: (crypto_alg_mod_lookup+0x118/0x210 <- crypto_probing_notify) return=0x0
	 stress-ng-af-al-26652 [048] d... 12159.931825: r_crypto_alg_mod_lookup_0: (crypto_alloc_tfm+0x41/0x100 <- crypto_alg_mod_lookup) return=0xfffffffffffffffe
	 stress-ng-af-al-26652 [048] d... 12159.931826: r_crypto_alloc_skcipher_0: (skcipher_bind+0xe/0x10 [algif_skcipher] <- crypto_alloc_skcipher) return=0xfffffffffffffffe
	 stress-ng-af-al-26652 [048] .... 12159.931827: sys_bind -> 0xfffffffffffffffe

So, stopping at those, as it doesn't seem to be a stress-ng issue.

Signed-off-by: Mauricio Faria de Oliveira <mfo@canonical.com>