stress-af-alg: add default configs to complement /proc/crypto list
The cryptographic algorithms listed in /proc/crypto are restricted
to those currently loaded, i.e., those the kernel is well aware of.

Previously, the hard-coded listed of algorithms allowed modules to
be loaded via the bind() syscall, when the crypto API looked it up,
via request_module().

Now, since we're restricting the list of cryptographic algorithms
that bind() is exposed to, to that list obtained from /proc/crypto,
that is restricted to currently loaded modules -- no more modules
are loaded, thus less cryptographic algorithms are being stressed.

Since the goal of stress-ng is to stress more of the system let's
revert back a little into the 'hard-coded list' previous approach,
but taking advantage of the /proc/crypto refactor/parsing.

This patch introduces this 'default configs' list, which is based
on information provided from /proc/crypto with the --af-alg-dump
option (from previous patch) via generated/included header file.

It then complements the dynamic list obtained from /proc/crypto,
so that more cryptographic algorithms make it to the total list,
and thus to bind(), which allows the kernel to request them up.

This patch changes the messages a little bit, because we're more
interested in knowing how many items are from /proc/crypto (less
if it's zero or not) and how many items are from the defconfigs.

And it removes the message about the number of algorithms to be
exercised, as it's not necessarily know before-hand, since some
might fail to bind(), thus not a precise claim even before this.

Before running with this patch:

  $ lsmod | wc -l
  126

  $ grep -c ^name /proc/crypto
  68

After:

  $ lsmod | wc -l
  193

  $ grep -c ^name /proc/crypto
  248

Signed-off-by: Mauricio Faria de Oliveira <mfo@canonical.com>