stress-af-alf: fix stack clobbering issue on arm64/armhf
For some reason as yet unknown we are seeing stack corruptions
when using the af-alg ciphers on ARM platforms. This can occur
with large buffers on the stack and signal handling.  To avoid
this, use heap allocated buffers and also add some allocation
slop to the buffers to try and avoid any buffer overflow heap
clobbering.  I've run this with valgrind and can't get any
corruption issues with this arrangement, so lets see how this
beds in with more exhaustive exercising on other architectures.

Signed-off-by: Colin Ian King <colin.king@canonical.com>