stress-close: add more open types, don't chown files if root, remove AF_APPLETALK
In a general improvement of this stressor add in some additional
ways to open a file using a tmp O_PATH file, a directory and a
shared memory object.  Also don't chown files if we are root as
we may be chowning /dev/zero, /dev/null etc and we want to avoid
this.  Finally, AF_APPLETALK on Linux is spamming the kernel log
and slowing the stressor down, so remove this.

Signed-off-by: Colin Ian King <colin.king@canonical.com>