stress-cyclic: handle differences in sched_attr sizes
Older kernels have a smaller sched_attr struct that does not provide
the SCHED_DEADLINE sched_util_min/max fields.  We can end up with
a mismatch when using an older kernel with a newer userspace and
-E2BIG is reported when trying to use SCHED_DEADLINE. In this case,
move to the next preferred scheduler which does not use these extra
fields and the kernel API does not return -E2BIG even when we use
a newer and larger sched_attr structure for non-SCHED_DEALINE
schedulers.

Signed-off-by: Colin Ian King <colin.king@canonical.com>