stress-shm: add some more shm exercising on shm fd
Add fchown on shm fd and move fallocate to exercise
the shm region a little more.  Also disable SIGCHLD so
that the exiting child does not interrupt any parent
system calls that are exercising the shared memory.
Add a page unmapping on the share memory segment on
the child to split pages.

Signed-off-by: Colin Ian King <colin.king@canonical.com>