stress-fcntl: exercise F_SETOWN, F_SETOWN_EX and F_GETOWN_EX more
It transpires that fcntl() is wrapped in libc and is using F_SETOWN_EX
for F_SETOWN on some systems so directly use the system call in these
cases.  Also define missing F_{SET|GET}_OWN types so that we can
exercise these if they are not defined.  Disable return checks for these
in case the kernel does not have the impemented yet.

Signed-off-by: Colin Ian King <colin.king@canonical.com>