set maximum file size based on off_t size
Currently the maximum file size was an arbitrary value. Fix this
by basing the upper limit to off_t.  If off_t is 32 bits we allow
(2^31)-1, of it's 64 bits then we allow (2^63)-1. off_t is signed
so we don't use the top bit.  When using the --maximize option
the maximum file size is set to be (2^31)-1 for 32 bit off_t or
(2^32)-1 for 64 bit. We don't want huge files for --maximize
to avoid taking a huge amount of time to write the initial test
files.

Signed-off-by: Colin Ian King <colin.king@canonical.com>