stress-sockfd: attempt to read from the fd passed over the socket
Add a bit more stress by peeking to see if there are any available
bytes to read on the fds passed over the socket. If there is data
then read 1 byte; this should in theory not block. Also use the
number of fd's sent over the socket as the bogo-ops metric rather
than each iteration of the send loop.

Signed-off-by: Colin Ian King <colin.king@canonical.com>