Ensure --aggressive mode terminates early (LP: #1858858)
Currently if --aggressive is used and the child processes have
terminated before the timeout the aggressive mode loop runs until
the timeout completes even when no more processes are running.
Add a waitpid NOHANG wait on each child process and if we get ESRCH
on all the child pids we can terminate the loop early.

Signed-off-by: Colin Ian King <colin.king@canonical.com>