Fix --random mode
The --random mode fails to start, this just gets "stuck" in startup:
  $ ./stress-ng --random 10

It gets stuck in busyloop in set_random_stressors(), trying to enable
randomly chosen stressors. But when set_random_stressors() is called, no
stressors are actually registered, so it's not able to enable anything.

Fix it by first registering all stressors before randomly selecting
which stressors we want to enable.

I believe the bug was introduced in commit:
9046db4eecaff5abdcc7725ba36734a4f382d1f6

Fixes: 9046db4eecaf ("Replace procs[] array with a doubly-linked list of proc_info_t")
Signed-off-by: Tommi Rantala <tommi.t.rantala@nokia.com>
Signed-off-by: Colin Ian King <colin.king@canonical.com>