core-helper: ensure data is null terminated
In the case where nothing is read because of an error null terminate
the data at the start of the buffer. In the case where the entire
buffer is filled, ensure last char is an eos of string char. In
the case where the buffer is partially filled, ensure eos char is
added to the end of the read data.

Signed-off-by: Colin Ian King <colin.king@canonical.com>