stress-ng: support sched deadline
There's an issue for current ' --sched deadline', that is we can't use this
policy actually in stressors, as the result of deadline parent can not be able
to fork() any child.  We can add the RESET_ON_FORK flag in monitor process
to make fork() possible. As the forked() child is CFS by default, we have to
reinit the child.

Besides, currently there's no parameters for the period/deadline/runtime for
deadline setup. So here add the sched-period, sched-runtime and sched-deadline
options to make deadline better use for users.

Signed-off-by: Chunyu Hu <chuhu@redhat.com>