stress-session: use wait4 if it is available
One of the untested kernel syscalls is wait4 to fetch rusage data,
so add a build time test for this, add a shim wrapper and call it
in the child session to wait for the parent.  We are not really
interested in the rusage stats, but it's worth calling this so
that exit in the kernel has to be exercised further to fetch this
data to pass to the waiter.

Signed-off-by: Colin Ian King <colin.king@canonical.com>