helper: make stress_warn_once simpler to use and remove flags
Currently the warn once mechanim required a flag to indicate the
uniqueness of the caller. This had only 32 slots and required keeping
a global set of flags.  Replace this with a version that takes the
filename and line number and hashes this into a 32 bit pattern that
is most probably unique and hides away all the details from the caller.
This implementation has space for 128 unique call points and that is
plenty for now.

Signed-off-by: Colin Ian King <colin.king@canonical.com>